Solve 347. Top K Frequent Elements by using a min heap
- By using a min heap, we can maintain a heap of size k as
opposed to a (max) heap of size n (and then determining the
k most frequent elements).

- We first simply add k elements to the min heap. After that, if
we attempt to add more elements to the heap,
1. We add the new element to the heap so that the size becomes
(k+1).
2. We then discard the minimum element (which can be done easily
for a min heap) and restore the heap property.

We add first and then pop as opposed to the reverse so that we
don't inadvertently pop off a value that was actually one of the
candidate maximums.

By repeatedly popping off the minimum values, we will be left
with the least minimum values i.e. maximum values at the end.
Since we start popping after adding k values to the heap, the
size of the final heap will be k with the k most frequent values.