1643 - Change the method of comparison based on value of k
If there are 100 elements and the 100th smallest element (k) is
asked for, that means that we want the largest element. In that
case, it is pointless to keep track of all 100 elements. We just
keep track of the largest element at each stage.

Based on the value of nCr(target_row + target_column, target_row)
: the upper bound for k given in the problem, if k goes beyond
the half point, use a normal comparison and therebyrepeatedly
pop off the minimum values.

If k is less than the half point, invert the comparison so that
max values become min and they are popped off from the heap.