235 - Add solution after learning about method from a discussion post
The idea is that if p and q are on opposite sides of the root,
then the root is the lowest common ancestor. If they're both on
the same side of the root, then we need to recurse further until
we get to a point where they are on either side of the root.

Link: https://leetcode.com/problems/lowest-common-ancestor-of-a-binary-search-tree/discuss/64963/3-lines-with-O(1)-space-1-Liners-Alternatives