1643 - Represent the paths using bits, stored as integer values
If the move is "H", assign a bit of 0. If the move is "V", assign
a bit of 1. We do this because H is lexicographically smaller
than V and correspondingly, 0 is less than 1 for binary.
e.g. A comparison between HHV and VHH becomes a comparison
between 001 (1) and 100 (4) wherein 001 is the lesser value.

Continue to wrap the path in an object since we need different
comparison strategies depending on the value of k and because
we can't change comparison strategies for integer values (the
paths are now stored as integers).

Take care of some binary stuff because leading 0s are ignored.
Consider a variable steps_away which denotes the number of steps
to reach the end from a cell located at (i, j). Use that to
generate the path values for a cell at (i, j).