1643 - Simplify things w.r.t the heap
Instead of wrapping the integer path in an object so as to
support different comparison strategies, use a "weight" variable
which is set to -1 if we want a max heap, or set to 1 if we want
a min heap. A min heap is needed when k is greater than half the
number of possible combinations.

When pushing into the heap, multiply the integer path with the
weight so that the product will serve as the key/comparison value
when pushing and popping from the heap.

The key value for the target cell doesn't matter since no other
cell can get to it using the current traversal strategy and
since the target cell is the starting point anyways - assign the
product to be 0.