1643 - Add a separate combinatorial-binary candidate solution
We find all the binary numbers of length target_row +
target_column that have target_row number of bits set in them
(target_row because that denotes the number of V moves or '1'
bits in the binary path).

Such numbers are not consecutive and so we iterate from a lower
bound of a binary number that has size target_row - 1 (if it
has this size, it cannot have target_row number of 1 bits).

We check the number of 1 bits of the generated numbers and
increment the count whenever we get a number with the required
number of 1 bits. Once the count reaches k, we generate a 'H, V'
path from the binary path that made the count equal to k.

The strategy to count the number of 1 bits was derived from the
page https://wiki.python.org/moin/BitManipulation - the section
'bitCount()'.

Note: This approach has a longer run-time than my original 5-mix
approach though and would still yield a Time Limit Exceeded
message.