986 - Use binary search to figure out the search window
With a left bisection/binary search in the second list, search
for the index of the end of f_intvl. This index will give the
separation point for all intervals in the second list that
start after f_intvl finishes. We then only need to search for any
intersections upto that point.

For the start value of the search window, use the previous search
index since that would be the possible first point of intersection
for f_intvl with all the intervals from the second list (draw it
out to get a sense of it).