Attempt 1 at 128. Longest Consecutive Sequence
Find the minimum value. Construct a set that contains differences
starting from 0 going up to (excluding) the length tof the list.
Any numbers that are consecutive with the minimum will have
differences starting from 0 (minimum value itself), 1, 2 ... so on.

Check the #differences before and after iterating over the list.
The difference between them will give the length of the longest
consecutive sequence.

18 / 70 test cases passed.

The problem with this is that the minimum value may not be a part
of the longest consecutive sequence. Also, even though differences
are being discarded from the set, the answer depends on
_consecutive_ differences that are discarded. If the discarded
differences are scattered in the set, they cannot be counted
together for the final answer.