Attempt 1 at 1319. Number of Operations to Make Network Connected
- Construct a (directed) graph from the connections. Don't need
the graph to be undirected.

- Run (multiple instances of) BFS on the possibly disconnected
graph of connections (hence the use of a set containing all the
computers). By running a BFS, we would create a tree that's using
the minimum number of edges to connect everything that's connected
to each other.

- The edges that are therefore left over, can be used to make the
entire graph connected. Each run of the BFS will generate a
connected component. We can rethink a component to be a single
node and we would therefore like to connect a bunch of isolated
notes. We can create a path between them. To form a path using
n nodes, we would need (n-1) edges.

- Check if #edges that are left over after the BFSes are
sufficient to create a path between the components/nodes.

6 / 36 test cases passed.