Partially solve 851. Loud and Rich: Time Limit Exceeded
- Build a directed graph using the input list 'richer' where an
edge is directed towards the richer person for each pair in 'richer'.

- For each person, run a DFS that traverses through it's progressively
richer neighbours and keep track of the lowest quietness along the way.

- For a person p, once the DFS is done, you will get a person q who
is the least quiet amongst all the people who definitely have equal
to or more money than p (q can be equal to p as well).

- Possible problem right now: There are repeated calculations
and additions to the stack happening. These can probably be
reduced.

71 / 86 test cases passed.