Partially solve 417. Pacific Atlantic Water Flow*
- Maintain another grid where each element is a 3-item list:
can cell reach Pacific Ocean, can cell reach Atlantic Ocean, is
cell visited.

- Run a regular DFS: check if the cell hasn't been visited, check
if the indices for the neighbours are valid, validate the heights
and only run a DFS on the neighbour if it hasn't been visited.

- Use the result from the neighbours to update the result for the
current cell. And in the end, return indices for which both oceans
are reachable.

108 / 113 test cases passed.

- Not sure what's wrong yet.