417 - Improve running time a bit (~1500 ms vs. ~3000 ms before)*
- When a DFS _originates_ at a cell A, set it's visited status to 1.
Otherwise, if cell A is encountered while a DFS that started from
any other cell B is running and A's status is 0, set A's status to
1 when you recurse into it but set it's status back to 0 when you
return from the recursive call.

- The idea being that a cell is considered "truly" traversed over
when a DFS starts from that cell itself. Any other time, we flip
it's status back to 0 so that other cells that want to use cell A
may be able to use it as well.