417 - Try out another modification*
- If the height of a cell and it's neighbour is equal, equate their
flow lists so that when one list gets updated, the other list gets
updated too. I believe this takes care of a 'weird' issue of two
cells having equal heights wherein it establishes a two way link
between the cells when running the DFS.

- Basically, it allows for a cleaner run of the DFS. And, a big
benefit now is that the solution is written using a regular DFS
where once a cell has been visited, we don't recurse into it again.
This should help cut down the time complexity and unnecessary
recursion and processing.

111 / 113 test cases passed.

- But still, the solution is failing for two test cases.