Solve 90. Subsets II
- For the numbers that are repeated, the kinds of subsets that they
yield are dependent on their quantities. For example, if 3 appears
four times, the subsets for 3 are: [3], [3, 3], [3, 3, 3] and
[3, 3, 3, 3].

- Figure out the subsets for the repeated numbers separately. Then,
for each of the subsets generated, extend the power set that exists
so far.

- Also, maintain a separate set 'visited' which will store the
repeated numbers from the list that we have encountered before.
These numbers have to be used only once so check their visit status.

- For the numbers that are not repeated, build the power set in the
same way that it was built for 78. Subsets.