300 - Try a more DP-like approach
- For any number A in the list, consider a slice of the list that
goes upto the number A's position in the list. In that slice,
determine the rightmost index of any number B such that B < A.

- The reason for that is that if A is to be a part of the sequence,
the number that precedes it must be smaller than it. And it would
make sense to get a smaller number that's at the righmost position
so that we get a sequence that's of the greatest size.

- Since B < A, a sequence containing B, A would be valid. Retrieve
the length of the longest increasing subsequence that was found
for number B and increment it by 1 since we are appending A to the
sequence. That establishes a DP-like relation between the numbers.

- If an index for a smaller, preceding number cannot be found, use
the answer for the number that lies at the previous index.

24 / 54 test cases passed.

- There seems to be a problem with this..