Solve 300. Longest Increasing Subsequence (!!)
- Continue to check the prefix array when finding the length of the
longest sequence upto a particular index.

- Add an additional condition while iterating through the prefix
array. Now, we are not only interested in the rightmost smaller
number. We want to find the rightmost smaller number that has the
greatest length value stored for it.

- Using the rightmost smaller number by itself won't necessarily lead
to the right answer. We also want to check the lengths of the
sequences yielded for each of the numbers in the prefix array.