300 - Attempt a solution to achieve O(nlogn) time complexity: Didn't work
- The idea is to use heaps and insert indices based on two priorities.
The higher priority is the length of sequence achievable with that
index and the lower priority is the index itself. We want the index
that is associated with the longest sequence and is the rightmost
possible index.

- Get the best possible index from the heap whilst checking that the
number at that index is less than the current number (otherwise, the
sequence wouldn't be ascending in nature). Store the indices that
didn't work out in a temporary storage area and then add them back
to the heap once an index has been determined.

- If a valid index is not found, the sequence length for that index
will be 1. But the list is initialized with values of 1, so no
extra steps need to be taken (A length of 1 means that the number
is by itself and forms a sequence of length 1).