39 - Try out a way to generate unique combinations from the get-go*
- This way, we don't have to sift through the combinations later to
get the unique ones.

- Use almost the same thing as before while generating combinations
but add an additional condition wherein we only add combinations
for a number x, if x > tgt - x. This will prevent repeated addition
of combinations.

- For example, for [2, 3, 5] and target = 5, calls will be made
with x = 2, tgt - x = 3 and x = 3, tgt - x = 2. This would lead to
repeated combinations. The extra condition ensures that we only add
the combinations from one of them.

- For the break condition, consider an example where tgt = 10 and 2
is a candidate. Then, the possible calls are (2, 8), (4, 6), (6, 4),
(8, 2) and (10, 0). The call (6, 4) and onwards satisfy the new
condition but we don't want all of them since they would give the
same combinations. Therefore, we break as soon as we get the first
x > tgt - x.

90 / 170 test cases passed.

- This solution should work I believe. Something is going wrong somewhere..