416 - Attempt an arithmetic-based solution
- If the total sum is odd, then it is not possible to partition
into 2 subsets with equal sum. The total sum has to be even for the
list to be partitioned.

- This translates to the question that we are looking for a selection
of numbers from the list that have a sum of (total sum // 2). The
leftover numbers will automatically have a sum of (total sum // 2)
as well.

- Sort the input numbers. Do a binary search on the list and keep
track of the sum of the list upto the calculated index. If the sum
upto that index is less than total sum // 2, continue with the
binary search on the righthand-side of the list.

- If the sum exceeds the target value, that means that the target
value cannot be made from a selection of the numbers.

64 / 116 test cases passed.

The problem with this is that this approach is selecting a
_contiguous_ selection of numbers for the sum. That may not lead
to the answer always.

For example, for [1, 1, 2, 2] (sorted), picking contiguous numbers
leads to an answer of False when in fact it is possible to pick 1
and 2 (non-contiguous).