Solve 39. Combination Sum (!!)*
- Sooo, interestingly enough, I arrived at the solution for this
while I was working on 416. Partition Equal Subset Sum. The current
solution is based on the work done in the commit 191b160.

- The idea is to iterate over the cartesian product of the candidates
and the sub-targets (starting from 0 and going up to target). Fix one
candidate and then come up with all the ways of achieving the range
of targets using only that candidate. Once the work for that candidate
is finished, continue using the other candidates and build off of the
work from the previous iterations.

- This was originally done in commit 191b160 using a 2D matrix where
the combinations were stored for results obtained using a sub-array
upto (and including) index i for a sub-target of j. Using a 2D matrix
here isn't necessary because at any time, only the results from the
previous i value are being used. It is possible to keep track of
everything by simply using a one-dimensional array that's only
dependent on values of j (the sub-targets).

- The answer array contains the different combinations for arriving
at the sub-targets. And each of those is filled progressively by
using one candidate at a time (whilst building off of the previous
results stored in there). Feel free to run the debugger and see a
run-through. A good example could be [13, 4, 2, 5, 3, 1] and target = 14.

- The sorting of the candidates isn't necessary it seems. Having
them sorted only yields that the answer array will be 'neat' and
there will be an ascending progression to the combinations.

- The running time is impressive too: 52 ms, ~95% !!