Solve 110. Balanced Binary Tree
- Since the definition of TreeNode cannot be changed to store the
height of a node, create an additional Height class. Once the
height of a node is determined, change the value of the node to
a Height object.

- If a node's value is of type Height, that means that it's height
has already been determined and the height can be returned right
away without any recalculations.

- I wonder though if this is okay to do since we ARE changing the
value of a node just because we're checking if the tree is balanced
or not..