110 - Apply a different approach to avoid mutating the tree
- The approach was learnt from the following LeetCode discussion post:
https://leetcode.com/problems/balanced-binary-tree/discuss/35691/The-bottom-up-O(N)-solution-would-be-better

- The idea is to run a DFS to determine the heights of the nodes.
While that DFS is running, we check if the difference between the
heights of any two subtrees exceeds 1. If that is the case, we
return -1 to denote that the tree is unbalanced.

- Once we hit that -1, we recursively return it up the call chain
and up to the original call. There, we check if the DFS yielded a
value of -1 or not. For trees that are balanced, the DFS would
simply yield the height of the tree (which would be a non-negative
value).