124 - Partially solve the problem*
- Keep track of 2 kinds of sums. One will be the best sum that can
be achieved by _using_/_including_ a node n. That means that the
possible paths would be n + right, n + left or n + right + left.

- We do this so that if an ancestor of n wants to use the "sub-answer"
obtained for the tree rooted at n, there is connectivity in the path.
There is connectivity in the path because the best sum at node n was
calculated by considering paths which ALL included n. This helps to
ensure that when an ancestor is using a sub-answer we don't
accidentally build sums using disconnected paths.

- But, the best sum possible may not include a node n. That's where the
second type of sum comes in. Use this other sum to keep track of the
maximum overall sum for the entire tree. This sum is not tied to or
conditioned to use a node n. It simply keeps track of the best overall sum.

- Update the max overall sum by considering the maximum sums from the left
and right children as well as the best sum obtained at the node n itself.

- Use a value of negative infinity for the max overall sum for empty nodes
since the non-empty nodes can have negative values and we don't want the
empty nodes to 'override' the sum in an unintended way.

68 / 94 test cases passed.

- Something is going wrong somewhere. The solution makes sense overall.