Solve 102. Binary Tree Level Order Traversal*
- Run a BFS on the tree. Use two variables to keep track of the
number of nodes at a level. Use one variable for the size of the
current level and one for what the size is going to be for the next
level.

- Start with a current level of 1 since the first level has the
root and there is only one node in the first level. Use the current
level as a counter. While the counter hasn't hit 0, any not-None
children that are encountered belong to the next level. Increment
the variable keeping track of the size for the next level.

- The counter for the current level goes down by 1 whenever we deque
a number from the queue. Once the counter reaches 0, it means that
we have explored all the nodes at that level. Update the current
level to the value for the next level and append a new list to
the 2D list traversal since the upcoming nodes will belong to a
new level.