572 - Try using Post Order Traversal*
- The problem with in-order traversal was that it can return false
positives (or false truthy answers). For example, for trees
[1, 2, 3] and [1, 2] the in-order traversals yield [2, 1, 3] and
[2, 1]. Now, the second list is a sub-list of the first but the
answer here is supposed to be that the second tree is _not_ a
subtree of the first tree.

- That is the case because the trees rooted at 1 must match
completely. They cannot match partially.

- So, I changed the traversal to Post Order so that the children
are taken into account before the node itself. Doing this yields
the correct answer for the above input.

116 / 182 test cases passed.

- This failed for an example where any random tree is given and the
second tree was just a node without any children. Finding that single
number in the first tree's traversal list is no guarantee that it's a
subtree because we also need to check for the presence of children
and the structure.

- Basically, a node with value 5 that is a leaf is not the same as
a node with value 5 that has a sub-tree rooted at it. The 5's will
both show up in the traversal lists but one is not a subtree of the other.

- Perhaps traversals may not be usable for this question..