Solve 572. Subtree of Another Tree*
Few pointers:
- The range needed to be fixed. It's wrong to use a step value that
is equal to the length of the traversal for the sub-root. We need
to progressively check the entire traversal for the root, with a
simple step value of 1.

- Post Order or Pre-Order traversal can be used here. In-Order
traversal cannot be used since a false truthy answer may be
returned due to partial matches.

- To take care of the problem from the previous commit about
example cases where the second tree is just a single node (other
problems are possible too but this was one that was observed/encountered),
when a node is None, append None to the traversal. This will allow
us to better keep track of the structure of a tree. This idea was
seen from the LeetCode discussion post here:
https://leetcode.com/problems/subtree-of-another-tree/discuss/102760/Easy-O(n)-java-solution-using-preorder-traversal
In particular, the line with the comment 'Appending # inorder
to handle same values but not subtree cases'.