98 - Partially solve the problem by recursively validating*
- Check if a node and it's children have appropriate values with
respect to each other. And then, check the trees rooted at the children.

72 / 80 test cases passed.

- The problem observed with this is that a child's value is checked
against it's parent but it may be invalid against one of it's
ancestors. For example, for a node 6 with left and right equal to
3, 7, the subtree at 6 would be valid. But if 6 was a right child
of 5, the current solution would say that the tree is a BST but that
is not the case since in the current tree, 3 is on the right side of
5, which is incorrect for a BST. We checked 3 against 6 but not against
it's ancestor 5.