98 - Write the solution by passing around only two values*
- Instead of appending and building a traversal list, which can
become as large as the entire tree, use a single number to keep
track of the current value. The new number that is yielded at each
step of the in-order traversal must be greater than the then
current value. If that is not the case at any point in time,
the numbers are not in ascending order and we can return False.

- Maintain another variable that stores the boolean results from
calls to the sub-trees.

- Maybe the solution with the whole traversal list in the previous
commit was a touch more readable but this isn't too bad either.