230 - Write optimized solution for 'Follow up' question using heaps*
- Since we're interested in finding the kth minimum, we need to get
rid of the maximums. But, the heapq module runs a heap as a min heap.
Therefore, to constantly get rid of the maximums, multiply the values
by (-1) before inserting them into the heap.

- This way, the larger a value is, the smaller it becomes when it goes
into the heap and therefore, it becomes more and more of a prime
candidate for a heap pop.

- Maintain a heap of size k to store the k smallest elements. Once
we've iterated over all the values, the value at the top of the heap
will be the kth smallest value.