105 - Attempt a solution that uses the preorder as a list of 'roots'
- Since a preorder traversal starts with the node and then moves onto
the children, the numbers in preorder denote the roots of various
subtrees. For these roots, check their indices in the inorder.

- The numbers to the left of the index obtained will denote the left
subtree for that number, and the numbers to the right of that index
will make up the right subtree. Use the roots from preorder to build
subtrees.

- This works out pretty nicely for a (proper) test case I created.
The only issue was when the preorder and inorder were the same,
which denotes a path-like tree that grows to the right. But I
handled that separately at the start of the function. Despite all
of this, only a few test cases are accepted as of right now.

9 / 203 test cases passed.

- I believe the problem is in the way the indices for preorder
are being selected for the recursive calls. Specifically, the issue
that I'm observing right now is with the index of preorder for
root.right. It seems like the index should not be 2.