211 - Partially solve the problem
- The trie here starts with a root node that doesn't hold any letter
value of it's own. The letters start appearing from the root's children.

- For the alphabet characters, traverse down the trie, looking for
the next character from word. If at any point any of the characters
is not found, return False.

- For the '.' character, any of the children of the current node
would be valid. Therefore, start a recursive search with those children
as roots and the portion of the word that comes after that '.' If the
remaining word is found using any of those children, return True, else
return False.

10 / 13 test cases passed.

- The solution makes sense but something seems to be going wrong
somewhere..