212 - Solution is working I believe but getting Time Limit Exceeded
- Perform a DFS starting from each cell on the grid. Keep track of
the path that's getting built along the way. Once a dead end is
encountered whilst traversing (dead end meaning that no other
neighbours can be visited from the current position), add the path
obtained thus far to the trie.

- Modify the insert method for the trie. Instead of only setting
the value to True for the last character of the word, set the value
to True for every character in the word. We do this because, in the
grid, it's possible to generate any sub-word along a DFS path.

- Find all the words that are possible from the grid and add them
to the trie. In the end, return those words from the input list
that are present in the trie.

- The definition of the Trie is taken from the solution I wrote for
208. Implement Trie (Prefix Tree).

21 / 62 test cases passed.

- Presumably, the time limit is getting exceeded because of traversing
over parts of the grid again and again. Some sort of caching of the
results/paths is probably needed.