212 - Add the input list of words to the trie instead of the words from the grid
- I read some LeetCode discussion posts and saw that the input list
of words were being added to the trie. And _then_ we would run
DFSes from every cell in the grid.

- Revert back to regular inserts for the trie wherein only the
last node gets a value of True, as opposed to setting the intermediate
nodes to True as well which I was doing before.

- The key part here would be that the structure of the trie would
determine where and when we recurse into a neighbour while running
a DFS. For example, if in the trie, there's a node 'p' that has
only one child 'a' and then in the grid, we're at a cell with value
'p' which has neighbours [h, k, a, b].

- Then, there would be no sense in visiting neighbours h, k and b since
they are not part of any word that starts with 'p' in the trie. It is
only worth visiting cell 'a' since that IS present in the trie after 'p'.

- This cuts down on the neighbours that we visit during a DFS to only
consider possibilities that are valid for the input list of words.

38 / 62 test cases passed.

- Still getting Time Limit Exceeded for some reason..