212 - Solution is accepted but is quite slow: 9200 ms, 10%
- Remove the search and startsWith methods for Trie since they are
not being used.

- So, I was reading on the discussion board that people are getting
a lot of Time Limit Exceeded even though their answers seem to be
correct. So, people suggested some modifications here and there.
Change the type for children to a regular dict instead of defaultdict.
Not sure if it's offering any real improvement in performance..

- In performDFS, if a node with a True value is encountered, append
the path to the list and also set the node's value to False so that
the same path isn't appended once more. I was simply using a set
before to deal with this, without having to set the node's value
to False but people in the discussion board did this. With this
modification, we don't need to use a set since a path will only
get added once.

- Change the way the visit status is being handled. Basically, we
want to avoid a node from being visited again if it's already part
of the DFS stack. People seemed to have done this by changing the
cell's value to a non-alphabet value. This way, the cell cannot
belong anywhere in the Trie and it will be skipped over. And, once
we're done DFSing into that node, we restore the value for that cell.
This achieves the same functionality of maintaining a visit status.

- Not sure if there's actually something wrong with the solution or
not. These 'modifications' aren't anything big and don't seem like
the 'proper' way to deal with this Time Limit Exceeded issue.