Solve 212. Word Search II: Add a real optimization of deleting a node once it's 'done'
- So, I think the modifications from the previous commit weren't
all that useful. Right now, the optimization is that once we're finished
with DFSing into a neighbour and we return from it, we check if it's
value is False and if it doesn't have any children.

- If that is the case, there will be nothing further for this node
to contribute and we can delete it. And since we will do this
recursively, we will be able to delete a lot of redundant nodes
and reduce the search space.

- I saw this optimization in one of the 'good' or faster solutions
from that runtime distribution graph that pops up after your solution
is accepted.

- Even though the deletion makes sense, I'm surprised that it made
such a big difference and was the reason why my solution(s) before
were timing out or were too slow.

- Runtime: 460 ms, 77%