417 - Solution is accepted but is slow: 1500 ms, 10%*
- Start a DFS from a cell if it has not been visited and add it to
visited. Alternatively, it's okay to use the membership test for
pacific_cells and atlantic_cells. Using either the visited set or
the ocean sets works.

- Once we're done visiting a neighbour within an already running DFS,
we remove it from visited so that the neighbour can be used by
other cells during that same DFS. This is the *KEY* difference between
the previous commit with 112 / 113 test cases passed and the current
commit with the accepted, albeit slow, solution.

- In the previous commit, when we weren't removing the neighbour after
we were finished DFSing into it, it ensured that any cell was visited
only once. This made for a fast DFS. But unfortunately, it wasn't enough
to give the complete solution (112 / 113 test cases passed).

- Now, by removing the neighbour from visited after the DFS into it, we
allow for more comprehensive DFSes on the grid, but we end up visiting
nodes more than ONCE. So, although we get greater completeness with this,
we end up traversing over nodes a lot more times, which leads to a slow
running time.

- I haven't been able to reconcile the two: fast, wrong and slow, correct.

- Don't remove the original cell that started the DFS within
pacificAtlantic_2ndPass() from visited. The idea being that since
that cell is the originator of it's own DFS (meaning that we didn't
arrive at this cell during a different cell's DFS), it will be able to
explore things completely freely and it can be considered "truly"
visited and so we don't have to remove it from the visited set.