417 - Check LeetCode discussion board for good approach*
- I finally decided to look at a solution for this problem. I'd
been doing this problem on my own for quite some time. So, the idea
is that we initiate a graph traversal FROM THE OCEAN.

- We start from the edges of the grid and move INWARDS to determine
how the water could flow. So, for this, we consider whether a neighbour
has _greater_ (or equal to) height than the current cell. We do this
because this tells us that when it does rain, the water would fall
from the neighbour onto the current cell and then it would go into
the ocean. So, we are looking for higher neighbours that would lead
to water falling on the current cell.

- We run a traversal from the 4 boundaries of the grid and we supply
the relevant ocean as a parameter to the traversal. For example, the
right boundary is touching the Atlantic Ocean and so we pass that.

- Run a DFS starting from the boundary and progressively add cells
to the respective ocean sets. Once a cell is added to the set, we
do not traverse over it again - this gives us a good DFS wherein
we don't repeat nodes/cells.

- In the end, return the intersection for the sets from the 2 oceans.