1976 - Add an implementation for Dijkstra's algorithm
- Check out the logic from the slides from Comp 251 (Lecture 11:
Single Source Shortest Path) and test the example that starts
from slide 36.

- Also, use https://stackabuse.com/dijkstras-algorithm-in-python/
to check implementation of the algorithm in Python.

- I did a debug run-through against the example starting from slide
36 and the procedure seemed to be good. I had to add one extra thing
on my own though. My observation was that a node can be added to the
heap more than once with different costs. Presumably, the version of
the node that gets added to the heap has progressively smaller and
smaller cost.

- But to avoid multiple iterations over that node, add a check after
the node is popped from the heap. If the node has already been visited,
then skip this iteration. The node that was popped probably belongs to
an older, larger weight version of the node.

- I came upon this observation when I noticed in the slides that the
weights for nodes were being updated AND the heap was getting updated
at the same time. While that makes sense on paper, it needs to be taken
care of programatically since the proper way of updating a heap is by
popping and pushing. You can't manipulate objects within the heap itself
and expect the heap to re-order itself based on that manipulation.

- Therefore, we get multiple instances of a node in the heap, with different
weights. And so, once that node has been popped and visited, we don't want to
visit it again.