1976 - Find #paths with shortest time to get to the destination
- Use Dijkstra's algorithm to determine the shortest path from the
source to the destination. Then, initiate a DFS from the source
and enumerate all the paths. At any moment, if the time for a path
exceeds the shortest time, do not recurse further down that path.

- Find all the paths for which the path time is equal to the
shortest path time. The 'path' variable isn't really needed. I was
simply using it to print which paths fit the bill. It can be removed.

16 / 54 test cases passed.

- Getting Time Limit Exceeded. Not sure what kind of optimization
is needed yet.