1976 - Start the traversal from the destination instead
- Change the approach and start traversing outwards from the
destination. Think of the shortest time it takes to get to the
destination as some amount of money p that you have when you start
your traversal from the destination.

- As we traverse outwards from the destination, whenever we use
an edge to get to another node, we must pay an amount that is equal
to the weight of the edge (the time value). So, any time our money
goes below 0, it is not possible for us to reach the source anymore.

- However, another optimization can be made here. We can keep track
of the minimum amount of money that we should have when we arrive at
a certain node x. For example, suppose that we had $10 left when we
got to a node. And we traversed deeper from this node and realized
that it was not possible for us to reach the source from node x with
$10. So, any other time if we come to node x with any value less than
$10, say $8, we don't need to traverse any further into node x since
we know that we weren't able to reach the source with $10 and so we
won't be able to get there with a smaller value like $8.

- This will allow us to not go down paths repeatedly and be able to
use the results from our previous computations. Start with a
placholder minimum value of -1. Any time we are about to recurse into
node x with 'money' p, store the number of paths we have found upto
that moment. After the DFS into node x with money p is done, check if
the number of paths changed.

- If the number of paths didn't change, that means that having this
combination of node and money didn't allow us to reach the source and
so for any future iterations, we want to have more money than p when we
intend to recurse into x. Continuously update the minimum money that
one should start with by considering the maximum of the old value and
the current money p that didn't lead to any addition of paths.

- Still getting Time Limit Exceeded unfortunately.

20 / 54 test cases passed.