1976 - Use results from Dijkstra's algorithm in the DFS
- When Dijkstra's algorithm is run from the source to all other
vertices, the values that get accumulated in min_times represent
the shortest time taken to reach all the vertices from the source.

- If we then use this data when we start the DFS from the
_destination_, we would already have the time boundaries for all
the nodes as a result of Dijkstra's algorithm.

- This is so because, in the DFS, when we reach a node x, we want
to know what's the least amount of 'money' p that we should have
before recursing into it. We have access to this information
through Dijkstra's output since the output there gives us the
shortest path with value p to get from node x to the source -
which can also be translated as, we need at least money p to be
able to travel from node x to the source.

- This allows us to not have to determine the time boundaries
while running the DFS. We already have access to it through
Dijkstra's algorithm.

20 / 54 test cases passed.

- Still getting Time Limit Exceeded though.