1976 - Attempt to optimize solution using dynamic programming
- So, in the previous commit(s), we were using the time boundary
to not recurse into a node x if the money left was less than the
minimum amount of money required to go from node x to the source.

- This helped deal with redundant money/time values that wouldn't
yield any additional paths. However, for queries at a node x that
_did_ have sufficient money to warrant a recursion into node x,
we were in a position where the same value might get calculated
many times.

- For example, suppose we arrive at x through a -> b -> c -> x
and want to know the #paths that are possible from x with money
of p. And suppose during another path exploration in the DFS, we
arrive at x through k -> l -> m -> x with the same money p. We
would end up doing the same work again in this situation.

- So, we want to cache the results as and when we calculate them
so that they can be reused. To this end, we would still have
the previous checks for a node not visited and money being
sufficient. But now, we add another check to see if a combination
of node x with money p was previously calculated or not.

- Use a defaultdict to achieve this in an easy manner without
having to worry about whether the keys exist or not. If a key
doesn't exist, False is returned which is the intended behaviour.
The structure of the dictionary is as follows: at the top level,
we have the node x. The value in the dictionary for a node x is
another dictionary wherein the key is the amount of money p that
we have at node x. And the value in _this_ dictionary is the
number of paths that are possible from node x to the source with
money p. So, the structure is {node: {money: #paths}}.

- This could be achieved with a 2D array as well but I thought
that we would have tons of empty spaces due to the high values
for money. This way, we only add the values we need.

- Continue using the results from Dijkstra's algorithm for the
shortest time between source and destination as well as the
time boundaries found for the nodes w.r.t. the source.

- Only update the #paths for (node x with money p) using one of
it's neighbours if the neighbour passes the money check. The money
check denotes whether the minimum amount of money was available or
not when wishing to recurse into a neighbour. If the money check
fails, there are no paths for the neighbour to contribute to node
x. Also, because we're using a defaultdict, redundant values would
be created in the dictionary if we tried accessing that neighbour.